Tue, Apr 23, 2024, 12:07 AM +05:30

+++ filter-less/helpers.c
@@ -7,7 +7,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int average = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
+            float average = (image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
 
             image[i][j].rgbtBlue = round(average);
             image[i][j].rgbtGreen = round(average);
